pyproject.toml
[tool.poetry]
name = "srmudp"
version = "0.0.2"
description = ""
authors = ["rid <rid_dim@gmx.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pycryptodome = "^3.22.0"
pyzmq = "^26.4.0"
pytest = "^8.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
